backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database and ORM
supabase==2.0.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client for external APIs
httpx>=0.24.0,<0.25.0
aiohttp==3.9.1

# Environment and Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# AI/ML Libraries
openai==1.3.7
requests==2.31.0

# Background Tasks
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.3
numpy==1.25.2

# Validation and Utilities
pydantic==2.5.0
email-validator==2.1.0

# Testing (for development)
pytest==7.4.3
pytest-asyncio==0.21.1

# Development tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1 